requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coverage==5.3.1
Django==3.1.5
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
httpie==2.3.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pygments==2.7.3
PyJWT==2.0.0
pylint==2.6.0
PySocks==1.7.1
pytz==2020.5
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
